Push main
I added DataPicker and Transportation. The DataPicker followed the default design throughout, and I found that it worked without errors in selecting dates.
I also changed the transportation from a radio button to a spinner. (for better visibility and clarity)
I made sure that all of this stored data is stored as a SharePreference and not as a DB.
I did this because I wanted to make sure that when the user presses the final save button, it would be saved to the DB.